feat: Enhance entity and player data structures for spatial partitioning
- Updated `EntityData` to include x, y, z coordinates.
- Introduced `PlayerData` structure to hold player position data.
- Modified `Input` structure to include a list of players and entity configuration.
- Refactored spatial partitioning logic to utilize `ValenceSpatialPartition`.
- Removed redundant spatial module and integrated its functionality into existing structures.
- Updated performance management to collect player data and integrate with Rust performance optimizations.
- Enhanced distance calculations and query methods in spatial structures for improved performance.

+++ rust/src/valence_integration/config.rs
@@ -49,7 +49,7 @@ pub fn load_config_from_toml(
                 entity_config.medium_radius = toml_config.throttling.medium_radius;
                 entity_config.close_rate = toml_config.throttling.close_rate;
                 entity_config.medium_rate = toml_config.throttling.medium_rate;
-                entity_config.far_rate = toml_config.throttling.far_rate;
+                entity_config.far_rate = toml_config.throttling.far_rate as f64;
             }
 
             // Update block config (assuming same as entity for now)
